Implemented Light and part of manager (testme!)
I am rather new to Python, and the Light class is the first one I’ve
ever written. There are bound to be bugs. I did try to write it to be
rather robust, though.
I can’t test on my Mac because pybluez is only compatible with Windows
and Linux.